Widgetize the Usage Card
All the behavior in app.js is now wrapped in a Marionette ItemView.
The interface is still being worked on, but basically you set the
el and model and called render when you're ready to display it.

I'm using marionette because it does a lot of the heavy lifting
managing memory and clean up, and providing useful helpers that
make the code much cleaner than if I use raw backbone.