Fix for Safari
For reasons I don't understand and can't duplicate on jsFiddle
On Safari 6.0.4, I am unable get a promise from a call to $.when
from invoking a function.

This causes the entire animation to stall while it waits for
the deferred to resolve. Returning a separate promise from the
deferred works, so I'm switching to that.

On Firefox and Chrome this works flawlessly but I am unable
to duplicate it on Safari. Safari also has some weird bugs around
the 'use strict' keyword, making the code undebuggable while the
flag is in effect :|