Refactor app.js - introduce state machine
The complexity of managing the states for the dashboard/graph/osd viz
was starting to get to be a bit much.

Introduce javascript-state-machine to manage the state transitions
between each mode. This adds it's own complexity but allows a clean
seperation into 3 events and 6 different edges that is more tractable
than managing the state in the DOM.