First pass rendering of host groupings
This a working but very inefficient version of the host groupings code.
It needs some serious refactoring because it's already a very
inefficient number of passes over the data set.

It performs ok because there are only 160 elements in the set, but it
needs some serious tuning. However visually it's roughly where it should
be and needs some other enhancements like a hover with the name of the
host for the grouping.