BUG: navigating in & out of graph fails to render
Because we close the view it cleans and removes the view and all it's
listeners. This is probably an abuse of close, and we should probably
set up the entire view from a new object.

Workaround, wrap render, so that it decorates the call to render.
Reinitialize the new host selector drop down. Move the delegate
events call to here instead of on ever call to show buttons.