Enhance Gauge Helpers - add icon events
Auto enhance gauge elements with event handers for:

	status:{ok|warn|fail}

Triggering these on a Marionette ItemView class will
cause statusIcons to change color and fontawesome icon class.
This depends on the following markup in the template

<div class='card-status-icon ok'>
  <span class="fa-stack fa-lg fa-2x">
    <i class="fa fa-square-o fa-stack-2x"></i>
    <i class="fa fa-check fa-stack-1x"></i>
  </span>
</div>

The stacked icon color is set by the outer class: ok, warn, fail.
The icon to replace is found by looking for fa-stack-1x.
Three classes are used for the icon: fa-check, fa-warning and
fa-exclamation-circle.

To make the ItemView change icon state send it an event.

e.g. inktank.PgView.trigger('status:warn')

This has a couple of advantages:

1. Uniform interface, all gauges will respond to a 'status:*' event in
the same way, but it decouples the logic for triggering that event.
2. follows the DRY principle, no need to cut and paste this to all the
widget classes.
3. gauges which don't support statusIcon will be safely ignored.