Fix #7092: Fix number formatting for small values
humanize.filesize attempts to add byte, kb and mb to the formatting.
This isn't going to work for small values as they're not file sizes.

Write a small shim to add 'K' and 'B' and use humanize.numerFormat to
truncate the digits to 1 decimal place. Use an '_' uTemplate to sanitize the
values.