Avoid a storm of callbacks during animation events
Each widget was invoking an optional callback to trigger an additional
event. This caused a mini storm of events based on the number of
widgets. Wrap the callback in a _.once so it only gets invoked once.
This was mostly visible when the widgets were reappearing and caused
unneceessary DOM reads/writes.