Issue #7859: remove target parameter from URL
Strange state behavior was happening because more than 1 parameter was
being set. Use the history push state APIs to change the URI and remove
the target parameter once we have navigated to the correct part of the
App.

Fix a race in GraphWall. GraphWall may not have fully rendered itself
by the time the App has navigated to the graph page. Guard it with a
Promise isReady().