Extract Functions
- decompose large function to classify and format minions into smaller functions to aid comprehension.

TODO

For small numbers of N this approach is ok, but if we end up having
thousands of hosts, it will need a rethink. Trying to do it within
1 pass rather than 6 or 7 would be the goal.

I also think it's possible to avoid bucketing the minions into
columns. That's a UI concern and can probably be achieved using mod
into a simple array that has been padded for the boundry.