Issue #8337 review uses of _hide()
Decorate all $hide instances instead of redirecting them. Angular-strap
only lets you intercept 2 of the possible paths into $hide, this is
often harmless, but if you actually need to augment the behavior,
clicking on backgrounds that are not 'static' will not use your handler.

Use _.wrap to decorate all $hide instances on the modal so we always get
called, even by the built in click handler.

Signed-off-by: Yan-Fa Li <yanfali@gmail.com>