Fix error callback issue in pool modification
When the user submit a pool modification with duplicate name
or reduced pg count,rest-api returns the validation error.
But the corresponding error callback is not getting called and
the form doesn't show the errors.

The issue is due to the interception of rest-api response to
change code 304 to code 200 and not properly handling other status
codes. This is now fixed by properly transmitting the errors.

Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>