Remove Google Analytics code
Both the login and dashboard pages include google analytics code. This
is commented out on the dashboard page so has no effect there, but is
live on the login page. This means any time someone tries to log in to
Calamari, their browser will hit Google's server with a bogus Google
Analytics site ID. This seems rather undesirable to me.

Signed-off-by: Tim Serong <tserong@suse.com>