Poll every 500ms to activate login button
There's another variant of issue #8352, where the form is initially
blank, but the user later types in a username which causes Firefox to
autofill the password field at this point in time, rather than on
initial pageload.  This does not trigger an input event, so the submit
button is never enabled.  To workaround this, I've changed setTimeout to
setInterval, so the page continually polls every 500ms.  This also means
we can remove the input event handlers, because the poll will pick up
the changed fields quickly anyway.

Signed-off-by: Tim Serong <tserong@suse.com>